We have a log!
Don't get to excited, this code is bad.  Ports is looking a bit
redundant and we have loads of types in the same file.  This method
of building everything in the same filecan be really helpful to get
things started but now it'sgetting a little hard to read.  the Ports
list is now doing a lesser job than the ships log. Lets clean that up!