Red Green together but that is some dirty code
OK we took a couple of step togther here and we got a green test but
in doing so we made the code pretty bad, there is loads of
duplication.  We not written very good object oriented code and there
is much we could put right but we still have a refactoring step to
rectify this.  Let see what we can put right and write it up in the
next commit.