Doesn't compile again!
Thought we were past this?  Well yes and no, I've slowed it down to
show how we use the tests to make design decisions.  Currently the
event source is just a list, this can't produce hydrated views. in
the case of the test we just wrote we made a design decision, that
procuding the view is the responsiblity of the event soource itself,
in this case a timespan.  This may change with time but that will be
easy as we have tests to make sure we don't break anything.