Test code problem resolved
In the last commit there were problems with the test code.  In reality
you wouldn't have commited this until the tests passed at least or may
be once you have refactored.  In this commit we have got a green board
but there is a horrible smell. When we ask the ships log for the
journey time we pass the ship to query on, this is clearly wrong.
Lets fix that up and may be add the final test.  How long does it take
to journey from Felixstowe to Porto de Santos in Brazil via Newark in
the USA.