First problem solved!
We set out to find out what the elapsed journey time was for the Ever
Given to travel from Felixstowe in the UK to Porta de Santos in Brazil
via Newark in the USA.  Solving it this way with an event source and
TDD we can scale the code, add more feature without the risk of a
combinatorial explosion.  We could add cargo, total layover times or
even average speeds per leg.  We have 8 production classes but only a
single test class because we have tested behaviour not implmentation
yet still maintained 100% test coverage.